Generalize even more and cleanup utils per env
